refactor: remove unused functions and modules
- Deleted my_unicorn/progress_manager.py and my_unicorn/utils/sha_utils.py - Removed related tests: test_progress_manager.py and test_sha_utils.py - Renamed
my_unicorn/utils/arch_utils.py to my_unicorn/utils/arch.py - Updated all imports and references from arch_utils to arch - Moved detect_checksum_hash_type into
SHAManager - Cleaned up unused code and test fixtures related to removed modules - Fix url installation error.
